.github/scripts/requirements.txt
pathlib==1.0.1
PyYAML==5.3.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
